requirements.txt
pynetworktables
opencv-python
openni
robotpy-cscore
tk
